pyproject.toml
[tool.poetry]
name = "pressnt"
version = "0.1.0"
description = "Longform writting social network"
authors = ["Daniel Valenzuela <dsvalenzuela@uc.cl>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.4"
Pillow = "^9.3.0"
django-markdownx = { "git" = "https://github.com/adi-/django-markdownx.git" }

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
djlint = "^1.19.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
